+++ .scalafmt.conf
@@ -1,4 +1,4 @@
 runner.dialect = scala3
 maxColumn = 90
 version = 3.8.6
-lineEndings=preserve
+lineEndings = preserve
