+++ fruits/src/main/scala/FruitEndpoint.scala
