+++ fruits/src/main/scala/FruitsHttpClient.scala
