+++ pillars/src/main/scala/CounterFactory.scala
