+++ pillars/src/main/scala/SwaggerMaker.scala
@@ -1,4 +1,3 @@
-
 import sttp.tapir.AnyEndpoint
 import sttp.tapir.server.ServerEndpoint
 import sttp.tapir.swagger.bundle.SwaggerInterpreter
@@ -6,5 +5,9 @@ import sttp.tapir.swagger.bundle.SwaggerInterpreter
 //noinspection ScalaUnusedSymbol
 object SwaggerMaker:
   private val interpreter = SwaggerInterpreter()
-  def fromEndpoints[F[_]](endpoints: List[AnyEndpoint], title: String, version: String): Seq[ServerEndpoint[Any, F]] =
+  def fromEndpoints[F[_]](
+      endpoints: List[AnyEndpoint],
+      title: String,
+      version: String
+  ): Seq[ServerEndpoint[Any, F]] =
     interpreter.fromEndpoints(endpoints, title, version)
